+++ webpack.common.js
@@ -20,9 +20,15 @@ const configureBabelLoader = (browserList) => {
         use: {
             loader: 'babel-loader',
             options: {
+                cacheDirectory: true,
                 presets: [
                     [
                         '@babel/preset-env', {
+                            modules: false,
+                            corejs: {
+                                version: 3,
+                                proposals: true
+                            },
                             useBuiltIns: 'usage',
                             targets: {
                                 browsers: browserList,
@@ -30,7 +36,10 @@ const configureBabelLoader = (browserList) => {
                         }
                     ],
                 ],
-                plugins: [],
+                plugins: [
+                    '@babel/plugin-syntax-dynamic-import',
+                    '@babel/plugin-transform-runtime',
+                ],
             },
         },
     };